Realizando melhorias no código com base na arquitetura limpa.
