+++ src/Core/Umio.API.Application/CasosDeUso/Clientes/Outputs/CriarClienteOutput.cs
