+++ src/Core/Umio.API.Entities/Entidades/Cliente.cs
@@ -80,5 +80,15 @@ namespace Umio.API.Entities.Entidades
             if (pontos.HasValue)
                 Pontos = pontos.Value;
         }
+
+        public static bool SenhaForte(string senha)
+        {
+            return !string.IsNullOrWhiteSpace(senha) &&
+            senha.Length >= 6 &&
+            senha.Any(char.IsUpper) &&
+            senha.Any(char.IsLower) &&
+            senha.Any(char.IsDigit) &&
+            senha.Any(c => !char.IsLetterOrDigit(c));
+        }
     }
 }
