+++ src/Core/Umio.API.Entities/Entidades/Enums/Provedor.cs
@@ -1,0 +1,9 @@
+namespace Umio.API.Entities.Entidades.Enums
+{
+    public enum Provedor
+    {
+        Umio,
+        Google,
+        Apple
+    }
+}
